// src/app/models/Imagenesgaleria.model.ts
import { StringifyOptions } from 'querystring';
import { NodeCompatibleEventEmitter } from 'rxjs/internal/observable/fromEvent';


export class ImagenesgaleriaModel{
    
    id: number;
    nombre: string;
    enlaces_id: string;
    rutapc: string;
    rutamovil: string;
    observaciones: string;
    posicion: string;
    tipos_id: string;
    titulo: string;
    subtitulo: string;
    texto: string;
    destino: string;
    enlace: string;
    enlaceppal: string;
    galeriafiltros_id: number;
    galeriafiltro: string;
    tipo: string;
    imagen: string;
    rutabase: string;
    descripcion: string;
    



}